+++ src/codegen.jl
@@ -1,9 +1,11 @@
 """
-    seek_connected(f, N::Int, head::DataType, tail::Tuple) -> (N, seq)
+    seek_connected(f, N::Int, head::DataType, tail::Tuple) -> (N, remainder)
 
 Recursively scan a tuple of `DataType` (split into its `head` and
-`tail`) to compute the uninterrupted sequence `seq` of adjacent
-operations (and its length `N`) where the predicate `f` is true.
+`tail`) to compute the length `N` of the uninterrupted sequence
+of adjacent operations where the predicate `f` is true.
+Additionally the `remainder` of the tuple (without that sequence)
+is also returned.
 """
 @inline function seek_connected(f, N::Int, head::Type{<:Operation}, tail::Tuple)
     if f(head)
@@ -32,51 +34,56 @@ end
 
 # --------------------------------------------------------------------
 
-function augment_impl(var_offset::Int, op_offset::Int, pipeline::Tuple)
-    augment_impl(var_offset, op_offset, first(pipeline), Base.tail(pipeline))
+function augment_impl(var_offset::Int, op_offset::Int, pipeline::Tuple, args...)
+    augment_impl(var_offset, op_offset, first(pipeline), Base.tail(pipeline), args...)
 end
 
-function augment_impl(var_offset::Int, op_offset::Int, pipeline::Tuple{})
+function augment_impl(var_offset::Int, op_offset::Int, pipeline::Tuple{}, args...)
     :($(Symbol(:img_, var_offset)))
 end
 
-function augment_impl(var_offset::Int, op_offset::Int, head, tail::NTuple{N,DataType}) where N
+function augment_impl(var_offset::Int, op_offset::Int, head::DataType, tail::NTuple{N,DataType}, avoid_eager = false) where N
     var_in  = Symbol(:img_, var_offset)
     var_out = Symbol(:img_, var_offset+1)
     if supports_lazy(head, tail)
-        num_affine, rest_affine = uses_affinemap(head, tail) ? seek_connected(uses_affinemap, 0, head, tail) : (0, nothing)
+        # If reached there are at least two adjacent lazy operations
+        num_affine, after_affine = uses_affinemap(head, tail) ? seek_connected(uses_affinemap, 0, head, tail) : (0, nothing)
         num_special, _ = seek_connected(x->(supports_permute(x)||supports_view(x)||supports_stepview(x)), 0, head, tail)
-        num_lazy, rest_lazy = seek_connected(supports_lazy, 0, head, tail)
+        num_lazy, after_lazy = seek_connected(supports_lazy, 0, head, tail)
         if num_special >= num_affine
             quote
                 $var_out = unroll_applylazy($(Expr(:tuple, (:(pipeline[$i]) for i in op_offset:op_offset+num_lazy-1)...)), $var_in)
-                $(augment_impl(var_offset+1, op_offset+num_lazy, rest_lazy))
+                $(augment_impl(var_offset+1, op_offset+num_lazy, after_lazy, avoid_eager))
             end
         else
             quote
                 $var_out = unroll_applyaffine($(Expr(:tuple, (:(pipeline[$i]) for i in op_offset:op_offset+num_affine-1)...)), $var_in)
-                $(augment_impl(var_offset+1, op_offset+num_affine, rest_affine))
+                $(augment_impl(var_offset+1, op_offset+num_affine, after_affine, avoid_eager))
             end
         end
     else
-        if length(tail) == 0 || supports_eager(head) || !supports_lazy(head)
+        # At most "head" is lazy (i.e. tail[1] is surely not).
+        # Unless "avoid_eager==true" we prefer using "applyeager" in
+        # this case because there is no neighbour synergy and we
+        # assume "applyeager" is more efficient.
+        if !supports_lazy(head) || (!avoid_eager && (length(tail) == 0 || supports_eager(head)))
             quote
                 $var_out = applyeager(pipeline[$op_offset], $var_in)
-                $(augment_impl(var_offset+1, op_offset+1, tail))
+                $(augment_impl(var_offset+1, op_offset+1, tail, avoid_eager))
             end
-        else # use lazy because there is no special eager implementation
+        else
             quote
-                $var_out = unroll_applylazy(pipeline[$op_offset], $var_in)
-                $(augment_impl(var_offset+1, op_offset+1, tail))
+                $var_out = applylazy(pipeline[$op_offset], $var_in)
+                $(augment_impl(var_offset+1, op_offset+1, tail, avoid_eager))
             end
         end
     end
 end
 
-function augment_impl(varname, pipeline::NTuple{N,DataType}) where N
+function augment_impl(varname::Symbol, pipeline::NTuple{N,DataType}, args...) where N
     quote
         img_1 = $varname
-        $(augment_impl(1, 1, pipeline))
+        $(augment_impl(1, 1, pipeline, args...))
     end
 end
 
@@ -102,8 +109,8 @@ end
 #       end
 #   end
 
-function augment_impl(pipeline::AbstractPipeline)
-    augment_impl(:input_image, map(typeof, operations(pipeline)))
+function augment_impl(pipeline::AbstractPipeline; avoid_eager = false)
+    augment_impl(:input_image, map(typeof, operations(pipeline)), avoid_eager)
 end
 
-augment_impl(op::Operation) = augment_impl((op,))
+augment_impl(op::Operation; kw...) = augment_impl((op,); kw...)